recombine split messages to fix monitor at low bauds
ioloop.py gets chunks of data for log and parses them for monitor
it was parsing each chunk which may or may not have full messages. at
low bauds, most messages were split and the monitor could go blank.

fix this by detecting the partial message, then keep it to combine with
the data in the next read